+++ gb/src/ppu/ppu.rs
@@ -544,7 +544,7 @@ impl Ppu {
         for (x, y, pixel, is_above_bg) in pixels {
             if self.is_cgb {
                 let obj_on_top = self.lcdc.bits() & 1 == 0;
-                let line_tile_attributes = self.line_tiles[x as usize];
+                let line_tile_attributes = self.line_tiles[x];
                 let bg_is_zero = line_tile_attributes.map_or(false, |pair| pair.0 == 0);
                 let bg_has_priority = line_tile_attributes
                     .map_or(false, |pair| pair.1 == BgToOamPriority::BgPriority);
